+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+1.1.2
+    fix bug: dragAndDrop.dragNode and dragAndDrop.overNode are not null when dragend event is triggered.
+
+1.1.1
+    methods:
+        remove method: getSearchResult()
+        search(keyword, fnMatch) returns matched nodes
+
 1.1.0
     props: remove prop 'fnMatch'
     methods: